provide infer() and merge() on Project
necessitated large refactor to avoid circular dependency
