+++ .gitignore
@@ -22,3 +22,4 @@ example_projects/*/derived/*
 example_projects/eg0-basic/pythinfer.yaml
 pythinfer.bak*.yaml
 demo-*.mov
+derived/*
