+++ src/pythinfer/__init__.py
@@ -1,5 +1,8 @@
-"""Initialise pythinfer package."""
+"""Initialise pythinfer package.
 
-from pythinfer.inout import Project
+Public API surface — exports the `Project` class from `pythinfer.api`
+"""
+
+from pythinfer.api import Project
 
 __all__ = ["Project"]
