+++ src/pythinfer/inout.py
@@ -1,36 +1,20 @@
-"""Input/output utilities for pythinfer package."""
+"""Input/output utilities for pythinfer package.
+
+This module contains I/O operations: dataset export, query loading, etc.
+
+The Project data model and project management functions (discovery, loading,
+creation) live in pythinfer.project. They are re-exported here for backward
+compatibility.
+"""
 
 import logging
 from collections.abc import Sequence
 from dataclasses import dataclass
 from pathlib import Path
 
-import yaml
-from pydantic import (
-    BaseModel,
-    ConfigDict,
-    Field,
-    ValidationInfo,
-    field_validator,
-    model_validator,
-)
-from rdflib import Dataset, Graph, Namespace, URIRef
+from rdflib import Dataset, Graph
 
 logger = logging.getLogger(__name__)
-
-# Base namespace for pythinfer graph identifiers and potentially other IRIs
-# Originally wanted to use a URN base (`urn:pythinfer:`) like so:
-# Format: urn:pythinfer:{project-name}:file:{relative-path}
-#     or: urn:pythinfer:{project-name}:inferences:{type}
-# However, parsing the TTL complained about no slash after colon etc.
-PYTHINFER_NS = Namespace("http://pythinfer.local/")
-
-PROJECT_FILE_NAME = "pythinfer.yaml"
-MAX_DISCOVERY_SEARCH_DEPTH = 10
-
-MERGED_FILESTEM = "0-merged"
-COMBINED_FULL_FILESTEM = "1-combined-full"
-INFERRED_WANTED_FILESTEM = "2-inferred-wanted"
 
 
 def export_dataset(
@@ -97,360 +81,6 @@ def export_dataset(
         logger.info("Exported %d triples to %s", len(dataset), fmt_output_file)
 
 
-def _get_sentinel_project_file() -> Path:
-    """Return a sentinel value used for the project file when generated by code."""
-    return Path.cwd() / "generated_by_code.nonexistent"
-
-def _empty_path_list() -> list[Path]:
-    """Give proper type to Field default below."""
-    return []
-
-class Project(BaseModel):
-    """Represents a pythinfer project configuration.
-
-    Attributes:
-        name: Name of the project.
-            specified directly in code.
-        focus: List of paths to focus data files. [Must be > 1]
-        reference: List of paths to reference data files (prev 'external'). [Optional]
-        path_self: Path to the project config file from which this came. May be none if
-
-    Note: path_self defaults to a sentinel value of "generated_by_code.nonexistent" in
-    the current working directory because this path is used to determine where the
-    output files go, and how the named graph IRIs are minted, and so this avoids a lot
-    of special-case code that would be necessary if we allowed path_self to be None.
-
-    Note: we manually implement hash and equality to exclude path_self, because it is
-    not relevant to the identity of the Project. If Pydantic implements a compare=False
-    option for fields in the future, we can switch to that.
-    https://github.com/pydantic/pydantic/discussions/6717
-
-    TODO: We could consider making the file lists sets because the order is irrelevant,
-    and repetition of the same file seems unwanted.
-
-    """
-
-    model_config = ConfigDict(
-        extra="forbid",  # This rejects unexpected keys
-        arbitrary_types_allowed=True,  # Allows Path objects
-    )
-
-
-    name: str
-    focus: list[Path] = Field(min_length=1)
-    reference: list[Path] = Field(default_factory=_empty_path_list)
-    owl_backend: str | None = None
-    sparql_inference: list[Path] | None = None
-    path_self: Path  = Field(default_factory=_get_sentinel_project_file)
-
-    def __eq__(self, other: object) -> bool:
-        """Compare Projects excluding path_self."""
-        if not isinstance(other, Project):
-            return NotImplemented
-        return (
-            self.name == other.name
-            and self.focus == other.focus
-            and self.reference == other.reference
-            and self.owl_backend == other.owl_backend
-            and self.sparql_inference == other.sparql_inference
-        )
-
-    def __hash__(self) -> int:
-        """Hash based on comparable fields only."""
-        return hash((
-            self.name,
-            tuple(self.focus),
-            tuple(self.reference),
-            self.owl_backend,
-            tuple(self.sparql_inference) if self.sparql_inference else None,
-        ))
-
-    @model_validator(mode="before")
-    @classmethod
-    def normalize_field_names(cls, data: dict[str, str]) -> dict[str, str]:
-        """Normalize field names to accept multiple spellings."""
-        # Map of alternative spellings to canonical field names
-        field_aliases = {
-            "data": "focus",
-            "local": "focus",
-            "paths_data": "focus",
-            "external-vocabs": "reference",
-            "external_vocabs": "reference",
-            "paths_vocab_ext": "reference",
-            "sparql-inference": "sparql_inference",
-            "sparql_inference": "sparql_inference",
-            "paths_sparql_inference": "sparql_inference",
-            "owl-backend": "owl_backend",
-        }
-
-        normalized: dict[str, str] = {}
-        for key, value in data.items():
-            # Use canonical name if it's an alias, otherwise keep original
-            canonical_key = field_aliases.get(key, key)
-            normalized[canonical_key] = value
-
-        return normalized
-
-    @field_validator(
-        "focus",
-        "reference",
-        "sparql_inference",
-        mode="before",
-    )
-    @classmethod
-    def convert_str_to_path(cls, v: list[str] | list[Path] | None) -> list[Path] | None:
-        """Convert string paths to Path objects."""
-        if v is None:
-            return None
-        return [Path(p) if isinstance(p, str) else p for p in v]
-
-    @field_validator(
-        "focus",
-        "reference",
-        "sparql_inference",
-        mode="after",
-    )
-    @classmethod
-    def resolve_relative_paths(
-        cls, v: list[Path] | None, info: ValidationInfo
-    ) -> list[Path] | None:
-        """Resolve relative paths against the project directory.
-
-        Uses the config_dir from validation context if available.
-        """
-        if v is None:
-            return None
-
-        config_dir = info.context.get("config_dir") if info.context else None
-        if not config_dir:
-            return v
-
-        resolved_paths: list[Path] = []
-        for path in v:
-            # Only resolve relative paths; absolute paths stay as-is
-            resolved_path = path if path.is_absolute() else config_dir / path
-            resolved_paths.append(resolved_path)
-        return resolved_paths
-
-    @classmethod
-    def from_yaml(cls, config_path: Path | str) -> "Project":
-        """Load project configuration from a YAML file.
-
-        Relative paths in the YAML file are resolved relative to the directory
-        containing the project file.
-        """
-        _config_path = Path(config_path).resolve()
-        config_dir = _config_path.parent
-
-        with _config_path.open() as f:
-            cfg = yaml.safe_load(f)
-
-        # TODO(robert): handle path patterns.
-        # TODO(robert): validate paths exist.
-
-        # Add path_self to the config dict before validation
-        cfg["path_self"] = _config_path
-        if "name" not in cfg:
-            cfg["name"] = _config_path.parent.stem
-
-        # Let Pydantic handle validation and field normalization
-        # Pass config_dir through context for path resolution in validators
-        return cls.model_validate(cfg, context={"config_dir": config_dir})
-
-    @classmethod
-    def discover(cls, start_path: Path | None = None) -> "Project":
-        """Discover the nearest project file and load it. Default to starting in cwd."""
-        return cls.from_yaml(discover_project(start_path or Path.cwd()))
-
-    def _path_to_yaml_str(self, path: Path) -> str:
-        """Convert a path to a string for YAML serialization.
-
-        If the path is relative to the project file's directory, store it
-        relative for better portability. Otherwise, store as absolute path.
-        """
-        resolved_path = path.resolve()
-        if self.path_self.stem.endswith("nonexistent"):
-            return str(path)
-        resolved_project_dir = self.path_self.resolve().parent
-        try:
-            # Try to make it relative to the project directory
-            rel_path = resolved_path.relative_to(resolved_project_dir)
-            return str(rel_path)
-        except ValueError:
-            # Path is not relative to project_dir, store as-is
-            return str(path)
-
-    def to_yaml_str(self) -> str:
-        """Serialize project configuration to a YAML string.
-
-        Paths are stored relative to the project file directory when possible,
-        for better portability.
-        """
-        cfg_dict: dict[str, object] = {
-            "name": self.name,
-            "focus": [self._path_to_yaml_str(p) for p in self.focus],
-        }
-        if self.reference:
-            cfg_dict["reference"] = [
-                self._path_to_yaml_str(p) for p in self.reference
-            ]
-        if self.owl_backend:
-            cfg_dict["owl_backend"] = self.owl_backend
-        if self.sparql_inference:
-            cfg_dict["sparql_inference"] = [
-                self._path_to_yaml_str(p) for p in self.sparql_inference
-            ]
-        return yaml.dump(cfg_dict)
-
-    def to_yaml(self, output_path: Path) -> None:
-        """Write project configuration to a YAML file."""
-        with output_path.open("w") as f:
-            f.write(self.to_yaml_str())
-
-    @property
-    def path_output(self) -> Path:
-        """Path to the output folder.
-
-        Includes the project file stem to avoid cache collisions when multiple
-        project files exist in the same directory.
-        """
-        return self.path_self.parent / "derived" / self.path_self.stem
-
-    @property
-    def paths_all_input(self) -> list[Path]:
-        """List of all input paths (focus + reference)."""
-        return self.focus + self.reference
-
-    @property
-    def paths_all(self) -> list[Path]:
-        """List of all paths (input + SPARQL inference) - cache checking."""
-        return self.paths_all_input + (self.sparql_inference or [])
-
-    @property
-    def namespace(self) -> Namespace:
-        """The IRI Namespace associated with this Project."""
-        # TODO: normalise name to be appropriate for an IRI.
-        return Namespace(PYTHINFER_NS[self.name] + "/")
-
-    @property
-    def provenance_gid(self) -> URIRef:
-        """The IRI to use for the provenance named graph for this Project."""
-        return self.namespace["provenance"]
-
-    def source_file_gid(self, file_path: Path) -> URIRef:
-        """Create a stable identifier for a source file's named graph.
-
-        Uses project name and relative path to create an IRI that is:
-        - Stable across re-parsing
-        - Portable within a project
-        - Informative about the source
-        - (no longer, because URNs don't currently work) Explicitly non-dereferenceable
-
-        Args:
-            file_path: Path to the source file
-
-        Returns:
-            IRI for the named graph, e.g.:
-            http://pythinfer.local/eg0-basic/file/basic-model.ttl
-
-        """
-        # Resolve both paths to their canonical form to handle symlinks and
-        # relative path differences that can occur across different environments
-        resolved_file_path = file_path.resolve()
-        resolved_project_parent = self.path_self.resolve().parent
-
-        try:
-            rel_path = resolved_file_path.relative_to(resolved_project_parent)
-        except ValueError:
-            # File is outside project directory; try to use the original path as-is
-            # to preserve the structure shown in the config file
-            try:
-                # Try with the unresolved path in case it has a meaningful structure
-                rel_path = file_path.relative_to(self.path_self.parent)
-            except ValueError:
-                # If that also fails, just use the file name
-                rel_path = resolved_file_path.name
-
-        # Note, to use a URN, we'd need to replace with colons for URN structure
-        # Use colons to maintain hierarchical structure in URN
-        return self.namespace[f"file/{rel_path}"]
-
-    def inference_gid(self, inference_type: str) -> URIRef:
-        """Create a stable identifier for an inference graph.
-
-        Args:
-            inference_type: Type of inference ('external', 'owl', or 'sparql')
-
-        Returns:
-            IRI for the inference graph, e.g.:
-            http://pythinfer.local/eg0-basic/inferences/owl
-
-        """
-        return self.namespace[f"inferences/{inference_type}"]
-
-def discover_project(start_path: Path, _current_depth: int = 0) -> Path:
-    """Discover a pythinfer project by searching for a config file.
-
-    Will recursively search parent directories until a config file is found or:
-    1. The root directory is reached.
-    2. A maximum search depth is reached (to avoid infinite recursion).
-    3. The `$HOME` directory is reached.
-
-    Args:
-        start_path: Path to start searching from.
-        _current_depth: Current search depth (used internally).
-
-    Returns:
-        Path to the discovered project config file
-
-    Raises:
-        FileNotFoundError if search reaches limit without discovering a project.
-
-    """
-    current_path = start_path.resolve()
-    config_path = current_path / PROJECT_FILE_NAME
-
-    # Positive case: config file found
-    if config_path.exists():
-        return config_path
-
-    # Negative cases: check search limits
-    msg = f"Search limit hit before finding project config (`{PROJECT_FILE_NAME}`)"
-    if current_path.parent == current_path:
-        raise FileNotFoundError(msg + ": reached root directory")
-    if _current_depth >= MAX_DISCOVERY_SEARCH_DEPTH:
-        raise FileNotFoundError(
-            msg + f": reached maximum search depth ({_current_depth})",
-        )
-    home_path = Path.home().resolve()
-    if current_path == home_path:
-        raise FileNotFoundError(msg + ": reached `$HOME` directory")
-
-    # Recurse to parent directory
-    return discover_project(current_path.parent, _current_depth + 1)
-
-
-def load_project(config_path: Path | None) -> Project:
-    """Load a pythinfer project specification from a YAML file.
-
-    The config file can either be specified directly, or discovered by searching.
-    If neither yield a result, a new project will be created.
-
-    Args:
-        config_path: Path to the config file, or None to trigger discovery.
-
-    """
-    try:
-        _config_path = config_path or discover_project(Path.cwd())
-        project = Project.from_yaml(_config_path)
-    except FileNotFoundError:
-        logger.info(
-            "⚠  No existing project found, creating new project in current directory",
-        )
-        project = create_project(Path.cwd())
-    return project
-
-
 @dataclass
 class Query:
     """Represents a query string more meaningfully than str."""
@@ -486,104 +116,3 @@ def load_sparql_inference_queries(query_files: Sequence[Path]) -> list[Query]:
             queries.append(q)
     return queries
 
-
-def create_project(
-    scan_directory: Path | None = None,
-    output_path: Path | None = None,
-    *,
-    force: bool = False,
-) -> Project:
-    """Create a new Project specification by scanning directory for RDF files.
-
-    Scans the specified directory (or current working directory) for RDF files
-    (with .ttl or .rdf extensions) and creates a new Project specification listing them.
-
-    This Project specification will be saved as `output_path` which defaults to
-    `pythinfer.yaml` in the scan_directory.
-
-    Subsidiary files such as SPARQL inference scripts are also sought and added to the
-    Project specification.
-
-    Args:
-        scan_directory: Directory to scan for RDF files. If None, uses current working
-            directory.
-        output_path: Path where the project file should be created.
-        force: Overwrite existing project file if it exists.
-
-    Returns:
-        Project specification object.
-
-    Side Effects:
-        Creates a new project configuration file at the specified output path.
-        May also create backup files if overwriting existing files.
-
-    Raises:
-        FileExistsError: If the output_path already exists.
-        FileNotFoundError: If no RDF files are found in the scan_directory or any of
-            its subdirectories.
-
-    """
-    _scan_dir = scan_directory or Path.cwd()
-    _output_path = Path(output_path or _scan_dir / PROJECT_FILE_NAME)
-
-    # Cowardly refuse to overwrite existing project file
-    if _output_path.exists():
-        if not force:
-            msg = f"Refusing to overwrite existing project file at `{_output_path}`"
-            raise FileExistsError(msg)
-        # Store backup of existing file with incrementing .bak suffix
-        for i in range(100):
-            backup_path = _output_path.with_suffix(f".bak{i}.yaml")
-            if not backup_path.exists():
-                break
-        else:
-            msg = "Too many backup files exist, cannot create new backup."
-            raise FileExistsError(msg)
-        _output_path.rename(backup_path)
-        logger.warning("⚠ Existing project file exists, backed up to `%s`", backup_path)
-
-    # Ensure output directory exists
-    _output_path.parent.mkdir(parents=True, exist_ok=True)
-
-    # Find all RDF files, excluding the 'derived' directory
-    rdf_files: list[Path] = []
-    for rdf_ext in ("*.ttl", "*.rdf"):
-        # Search recursively but exclude 'derived' directory
-        for rdf_file in _scan_dir.rglob(rdf_ext):
-            # Skip files in 'derived' directory
-            if "derived" in rdf_file.parts:
-                continue
-            # Store relative paths from scan directory
-            rel_path = rdf_file.relative_to(_scan_dir)
-            rdf_files.append(rel_path)
-
-    if not rdf_files:
-        msg = f"No RDF files found in directory `{_scan_dir}` to create project."
-        raise FileNotFoundError(msg)
-
-    # Sort for consistent output
-    rdf_files.sort()
-
-    # Load SPARQL inference queries from the 'queries' directory
-    sparql_query_files: list[Path] = []
-    for query_file in _scan_dir.rglob("infer*.rq"):
-        rel_path = query_file.relative_to(_scan_dir)
-        sparql_query_files.append(rel_path)
-
-    # Sort for consistent output
-    sparql_query_files.sort()
-
-    # Create project configuration
-    project_config = Project(
-        name=_scan_dir.name,
-        path_self=_output_path,
-        focus=rdf_files,
-        reference=[],
-        sparql_inference=sparql_query_files,
-    )
-
-    project_config.to_yaml(_output_path)
-
-    logger.info("✅ Created new project file at `%s`", _output_path)
-
-    return project_config
