+++ tests/unit/test_python_api.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 
-from pythinfer.inout import Project
+from pythinfer.project import ProjectSpec
 
 
 class TestProjectInstantiation:
@@ -10,7 +10,7 @@ class TestProjectInstantiation:
 
     def test_direct_instantiation(self) -> None:
         """Test that we can directly instantiate."""
-        proj = Project(name="test123", focus=[Path("blah.ttl")])
+        proj = ProjectSpec(name="test123", focus=[Path("blah.ttl")])
         assert proj.name == "test123"
         assert proj.path_self.stem == "generated_by_code"
         assert proj.path_self.suffix == ".nonexistent"
